add logic to a seperate file
